🚚 F03: APP> rename path files
