+++ src/components/pages/about/AboutPage.js
@@ -2,7 +2,7 @@ import Banner from "../../reusable-ui/Banner";
 import Collapse from "../../reusable-ui/Collapse";
 import Footer from "../../reusable-ui/Footer";
 import NavBar from "../../reusable-ui/NavBar";
-import bannerAbout from "../about/images/bannerAbout.png";
+import BannerAbout from "../../../assets/images/BannerAbout.svg";
 import styled from "styled-components";
 import infoAbout from "../../../assets/Datas/infoAbout.json";
 
@@ -10,7 +10,7 @@ export default function About() {
   return (
     <AboutStyled>
       <NavBar />
-      <Banner ClassName="banner" image={bannerAbout} />
+      <Banner ClassName="banner" image={BannerAbout} />
       {infoAbout.map((info) => (
         <Collapse title={info.title} texte={info.description} />
       ))}
